+++ CHANGELOG.md
@@ -6,6 +6,17 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## [Checkpoint] - 2025-12-23 15:54:32
+
+### Changed Files
+- `src/services/recruiting/checklistService.ts`
+- `src/services/recruiting/repositories/RecruitPhaseProgressRepository.ts`
+
+### Statistics
+```
+ 2 files changed, 24 insertions(+), 1 deletion(-)
+```
+
 ## [Checkpoint] - 2025-12-23 15:47:15
 
 ### Changed Files